# RGW/STS: honor configured limits when updating max session duration